pyproject.toml


[tool.poetry]
name = "food_truck_tracker"
version = "0.1.0"
description = "A sample Python project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "3.10.14"
requests = "^2.26.0"
Django = "5.0.6"
djangorestframework = "3.15.1"



[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
